put handle logout logic into Navbar.jsx
