+++ src/components/Navbar.jsx
@@ -1,6 +1,7 @@
 import { useState } from "react"
+import { logout } from '../api/security'
 
-export default function Navbar ({ handleLogout }) {
+export default function Navbar ({ setIsLogin, setMessages }) {
   const [isBlackTheme, setIsBlackTheme] = useState(true)
 
   // Change of theme
@@ -14,6 +15,38 @@ export default function Navbar ({ handleLogout }) {
     ) // We delete the old theme
 
     setIsBlackTheme(localStorage.theme !== 'light') // We change the icon of the theme
+  }
+
+  // Logout
+  const handleLogout = () => {
+    logout()
+      .end(function (error, res) {
+        // We process the data received
+
+        if (error) throw error
+
+        const data = JSON.parse(res.text) // The JSON data is converted into data that can be understood by Javascript
+        if (data.status == 200) {
+          // If the request has worked then
+
+          setIsLogin(false) // Display of the modal
+          setMessages([]) // The message table is made empty
+        } else {
+          setMessages(messages => {
+            // Sending an error message
+            return [
+              {
+                content:
+                  'Une erreur est survenue lors de la requête. Merci de contacter un administrateur.',
+                classes: ['text-red-500'],
+                background: true,
+                isQuestion: false
+              },
+              ...messages
+            ]
+          })
+        }
+      })
   }
 
   return (
